requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
keyring==23.5.0
numpy==1.23.5
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.6
sacremoses==0.0.53
scipy==1.10.1
triton==2.0.0
uvicorn==0.20.0
zipp==1.0.0
openai==0.27.4
pymongo==3.13.0
pandas==2.0.2
dnspython==2.3.0